chore: Some minor refactoring of ECC package into a sub package.
